+++ alembic/versions/009_add_playlists.py
@@ -8,8 +8,9 @@ Create Date: 2026-03-27
 from collections.abc import Sequence
 
 import sqlalchemy as sa
-from alembic import op
 from sqlalchemy.dialects.postgresql import UUID
+
+from alembic import op
 
 revision: str = "009"
 down_revision: str | None = "008"